Generate a plugin-compatible notification directly
Instead of creating a notification that launches the main activity, and in
turn, launches the the plugin-compatible notification, launch the
plugin-compatible notification directly.

This ensures that:
- we only have one notification, not two
- clicking on the notification goes directly to the status screen
- the notification is removed when it is clicked

This required some careful investigation:
- https://github.com/e-mission/e-mission-docs/issues/680#issuecomment-1031083191
- https://github.com/e-mission/e-mission-docs/issues/680#issuecomment-1031097154

And a related UI-only change
https://github.com/e-mission/e-mission-docs/issues/680#issuecomment-1031912649
